+++ CDTDatastore/CDTReplicator/CDTReplicator.h
@@ -16,6 +16,7 @@
 #import <Foundation/Foundation.h>
 
 #import "CDTReplicatorDelegate.h"
+#import "CDTNSURLSessionConfigurationDelegate.h"
 
 @class CDTDatastore;
 @class TDReplicatorManager;
@@ -80,6 +81,7 @@ typedef NS_ENUM(NSInteger, CDTReplicatorState) {
     CDTReplicatorStateError
 };
 
+
 /**
  A CDTReplicator instance represents a replication job.
 
@@ -127,6 +129,19 @@ typedef NS_ENUM(NSInteger, CDTReplicatorState) {
 @property (nonatomic, weak) NSObject<CDTReplicatorDelegate> *delegate;
 
 /**
+ * Set the delegate for handling customisation of the NSURLSession
+ * used during replication.
+ *
+ * This allows the setting of specific options on the NSURLSessionConfiguration
+ * to control the replication - e.g. replication only when on Wifi would be
+ * achieved by setting the NSURLSessionConfiguration's allowsCellularAccess
+ * attribute to 'NO'.
+ *
+ * @see CDTNSURLSessionConfigurationDelegate
+ */
+@property (nonatomic, weak) NSObject<CDTNSURLSessionConfigurationDelegate> *sessionConfigDelegate;
+
+/**
  Returns true if the state is `CDTReplicatorStatePending`, `CDTReplicatorStateStarted` or
  `CDTReplicatorStateStopping`.
 
@@ -146,6 +161,7 @@ typedef NS_ENUM(NSInteger, CDTReplicatorState) {
  */
 -(id)initWithTDReplicatorManager:(TDReplicatorManager*)replicatorManager
                      replication:(CDTAbstractReplication*)replication
+           sessionConfigDelegate:(NSObject<CDTNSURLSessionConfigurationDelegate> *)delegate
                            error:(NSError * __autoreleasing*)error;
 
 /*
@@ -170,17 +186,28 @@ typedef NS_ENUM(NSInteger, CDTReplicatorState) {
  * The replication will continue until the replication is caught up with the source database;
  * that is, until there are no current changes to replicate.
  *
- * -startWithError can be called from any thread and will immediately return. It queues its work
- * on a separate replication thread. The methods on the CDTReplicatorDelegate
+ * -startWithTaskGroup:error: can be called from any thread and will immediately return. It queues 
+ * its work on a separate replication thread. The methods on the CDTReplicatorDelegate
  * may be called from the background threads.
  *
  * A given CDTReplicator instance cannot be reused. Calling this method more than once will
  * return NO. Only when in `CDTReplicatorStatePending` will replication.
  *
+ * @param taskGroup an optional dispatch_group_t allowing clients to wait for the completion
+ *                  of all replication tasks in the taskGroup before proceeding.
  * @param error the error describing a failure.
  * @return YES or NO depending on success.
  *
  * @see CDTReplicatorState
+ */
+- (BOOL)startWithTaskGroup:(nullable dispatch_group_t)taskGroup error:(NSError *__autoreleasing *)error;
+
+/**
+ * Starts a replication.
+ *
+ * Calling this is the same as calling startWithTaskGroup:error: with a nil taskGroup.
+ *
+ * @see startWithTaskGroup:error:
  */
 - (BOOL)startWithError:(NSError *__autoreleasing *)error;
 
@@ -213,7 +240,7 @@ typedef NS_ENUM(NSInteger, CDTReplicatorState) {
  * to stop the replication before it starts. If it cannot stop the replication, this method
  * will return NO and replication will start and progress as normal.
  *
- * If -startWithError was never called, this method will move the state from
+ * If -startWithTaskGroup:error: was never called, this method will move the state from
  * CDTRelicatorStatePending to CDTReplicatorStateStopped, inform the delegate and return YES.
  *
  *