+++ CDTDatastore/touchdb/ChangeTracker/TDURLConnectionChangeTracker.h
@@ -8,7 +8,7 @@
 
 #import "TDChangeTracker.h"
 
-@interface TDURLConnectionChangeTracker : TDChangeTracker <NSURLSessionTaskDelegate>
+@interface TDURLConnectionChangeTracker : TDChangeTracker <NSURLSessionTaskDelegate, CDTURLSessionTaskDelegate>
 
 // used only for testing and debugging. counts the total number of retry attempts and
 // is not reset to zero with each separate request (unlike TDChangeTracker retryCount).